+++ src/bin/rew/pattern/help.rs
@@ -53,9 +53,7 @@ pub const FILTERS: &str = indoc! {r#"
   `s:X`    Remove match of `X`          (`S` = all occurences)
 
   `@:X1:Y1:...:Xn:Yn:D`  Output `Yi` for first match of `Xi`, `D` for no match
-
-  `0`, `1`, `2`, ...     Capture group of a global regex
-  `$0`, `$1`, `$2`, ...  Capture group inside `s/S/@`
+  `$0`, `$1`, `$2`, ...      Capture group of a global regex or `s/S/@` regex
 
 # FORMATTING